+++ CHANGELOG.md
@@ -11,6 +11,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
 ### Fixed
 
 - irc: limit size of data received from the server to prevent memory exhaustion
+- relay: limit size of received websocket frame and HTTP body to prevent memory exhaustion
 
 ## Version 4.9.1 (2026-05-31)